requirements_trt.txt
onnx
onnxruntime-gpu
tensorrt-cu12
tensorrt-lean-cu12
tensorrt-dispatch-cu12
